% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_pairwise.R
\name{do_pairwise}
\alias{do_pairwise}
\title{Do pairwise comparisons from a model object}
\usage{
do_pairwise(model, by, correction, ci = 0.95)
}
\arguments{
\item{model}{model object}

\item{by}{variable name over which to calculate the marginal means}

\item{correction}{correction method. (e.g. "bonferroni", "holm", "BH")}

\item{ci}{width of confidence interval. Defaults to 0.95.}
}
\value{
data frame of pairwise comparisons, adjusted p-values, and confidence intervals.
}
\description{
This function pulls emmeans::emmeans() and emmeans::pairs() to do pairwise
comparisons with user-chosen multiple testing correction method. In addition
to providing the means and p-values, it also produces confidence intervals based
on the standard error.
}
