% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emmeans.R
\name{get_emmeans}
\alias{get_emmeans}
\title{Retrieve marginal means}
\usage{
get_emmeans(model, by)
}
\arguments{
\item{model}{model object}

\item{by}{variable name over which to calculate the marginal means}
}
\value{
data frame
}
\description{
This function pulls emmeans::emmeans() into a pipeline that returns a dataframe
containing the marginal means.
}
