% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.bc.av.R
\name{optim.bc.av}
\alias{optim.bc.av}
\title{Find optimal lambda for BoxCox transformation}
\usage{
optim.bc.av(
  formula,
  groups = 1,
  data,
  K = 3,
  steps = 500,
  tol = 0.5,
  start = "gq",
  EMdev.change = 1e-04,
  find.in.range = c(-3, 3),
  s = 60,
  plot.opt = 3,
  verbose = FALSE,
  noformat = FALSE,
  ...
)
}
\description{
This function is a slightly modified version of the boxcoxmix::optim.boxcox()
function -- I added a progress bar. See ?boxcoxmix::optim.boxcox() for more information.
}
