% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar_randint.R
\name{univar_randint}
\alias{univar_randint}
\title{Build a univariate LMM with random intercept}
\usage{
univar_randint(y, x, gp, data)
}
\arguments{
\item{y}{response variable name}

\item{x}{single predictor variable name}

\item{gp}{random slope variable name}

\item{data}{name of the dataframe being used}
}
\value{
model object
}
\description{
This function will build a univariate LMM with random intercept
for a response variable that needs Box-Cox transformation. The function will
find the optimal lambda using the optim.bc.av function (adapted from the
boxcoxmix package), and produce a model using the lme4 package.
}
